+++ src/platform/macos.rs
@@ -33,18 +33,6 @@ pub trait WindowExtMacOS {
     /// space or taking control over the entire monitor.
     fn set_simple_fullscreen(&self, fullscreen: bool) -> bool;
 
-    /// Returns whether or not the window is occluded.
-    ///
-    /// Occlusion differs from window visibility as an occluded window
-    /// can be occluded from the screen surface for reasons such as being closed,
-    /// minimised, not set visible, or behind another window.
-    ///
-    /// Alternatively the event [`WindowEvent::Occluded`]
-    /// can be used to track the occlusion state of the window.
-    ///
-    /// [`WindowEvent::Occluded`]: crate::event::WindowEvent::Occluded
-    fn is_occluded(&self) -> bool;
-
     /// Returns whether or not the window has shadow.
     fn has_shadow(&self) -> bool;
 
@@ -71,11 +59,6 @@ impl WindowExtMacOS for Window {
     #[inline]
     fn set_simple_fullscreen(&self, fullscreen: bool) -> bool {
         self.window.set_simple_fullscreen(fullscreen)
-    }
-
-    #[inline]
-    fn is_occluded(&self) -> bool {
-        self.window.is_occluded()
     }
 
     #[inline]